+++ client/index.js
@@ -53,4 +53,3 @@ if (module.hot) {
   module.hot.accept('./index.js');
   module.hot.accept('../app', renderApp);
 }
-
