--- routes/app.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.get('/', function(req, res, next) {
-  res.render('app', { title: 'Shopify Node App' });
-});
-
-module.exports = router;
