+++ server.js
@@ -1,8 +1,8 @@
 const fs = require('fs');
 const express = require('express');
+const session = require('express-session');
 const path = require('path');
 const logger = require('morgan');
-const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 
 const webpack = require('webpack');
@@ -10,7 +10,18 @@ const webpackMiddleware = require('webpack-dev-middleware');
 const webpackHotMiddleware = require('webpack-hot-middleware');
 
 const config = require('./webpack.config.js');
-const appRoute = require('./routes/app');
+const shopifyAuth = require('./routes/shopifyAuth');
+
+const shopifyConfig = {
+  host: 'https://kevin-shopifyapps.fwd.wf',
+  apiKey: 'e83fbb0a19687cc6701eeb3ccdccb38c',
+  secret: '01a9b5f3e112808a868308c2bbf33dfe',
+  scope: ['write_orders, write_products'],
+  afterAuth: (req, res) => {
+    // do stuff like register webhooks
+    return res.redirect('/app')
+  }
+}
 
 const app = express();
 const isDeveloping = process.env.NODE_ENV !== 'production';
@@ -20,10 +31,13 @@ app.set('view engine', 'ejs');
 app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
-app.use(cookieParser());
+app.use(session({ secret: 'keyboard cat', resave: true, saveUninitialized: false }));
 
-// Put auth routes here!
-// app.use('/auth', ShopifyAuth);
+app.get('/',  function(req, res) {
+  res.render('install')
+})
+
+app.use('/auth/shopify', shopifyAuth(shopifyConfig));
 
 // Run webpack hot reloading in dev
 if (isDeveloping) {
@@ -49,7 +63,13 @@ if (isDeveloping) {
   app.use(express.static(__dirname + '/assets'))
 }
 
-app.get('*', appRoute);
+app.get('/app', function(req, res) {
+  if (req.session.access_token) {
+    res.render('app', { title: 'Shopify Node App', apiKey: shopifyConfig.apiKey, shop: req.session.shop });
+  } else {
+    res.redirect('/');
+  }
+});
 
 app.use(function(req, res, next) {
   var err = new Error('Not Found');
