+++ views/app.ejs
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 <!DOCTYPE html>
 <html>
   <head>
@@ -8,30 +7,21 @@
     <link rel='stylesheet' href='/stylesheets/style.css' />
   </head>
   <body>
+    <script src="https://cdn.shopify.com/s/assets/external/app.js"></script>
+
+    <script>
+      window.apiKey = "<%= apiKey %>"
+      window.shopOrigin = "https://<%= shop %>"
+    </script>
+
+    <script type="text/javascript">
+      ShopifyApp.init({
+        apiKey: apiKey,
+        shopOrigin: shopOrigin
+      });
+    </script>
+
     <div id="root"></div>
     <script src="/assets/main.js"></script>
   </body>
 </html>
-||||||| merged common ancestors
-=======
-<!DOCTYPE html>
-<html>
-  <head>
-    <title><%= title %></title>
-    <link rel='stylesheet' href='/stylesheets/style.css' />
-  </head>
-  <body>
-    <h1><%= title %></h1>
-    <p>Welcome to <%= title %></p>
-
-    <script src="https://cdn.shopify.com/s/assets/external/app.js"></script>
-
-    <script type="text/javascript">
-      ShopifyApp.init({
-        apiKey: "<%= apiKey %>",
-        shopOrigin: "https://<%= shop %>"
-      });
-    </script>
-  </body>
-</html>
->>>>>>> explored passport and a route factory
