--- views/index.ejs
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title><%= title %></title>
-    <link rel='stylesheet' href='/stylesheets/style.css' />
-  </head>
-  <body>
-    <header>
-      <h1>Shopify Node App – Installation</h1>
-      <p class="subhead">
-        <label for="shop">Enter your shop domain to log in or install this app.</label>
-      </p>
-    </header>
-
-    <div class="container__form">
-      <form method="GET" action="auth/shopify">
-        <input type="text" name="shop" id="shop" placeholder="example.myshopify.com"/>
-        <button type="submit">Install</button>
-      </form>
-    </div>
-  </body>
-</html>
